test: Trade 단위 테스트 작성
